CPP_06

+++ CPP_06/ex01/Serializer.hpp
@@ -4,7 +4,7 @@
 #include <iostream>
 
 struct Data{
-    int  x;
+    uintptr_t  x;
 };
 
 class Serializer
